# backend/stay-service/openapi.yml
openapi: 3.0.3
info:
  title: stay-service API
  description: >
    Manages the main stay lifecycle (check-in / check-out) of the parking system.
    Orchestrates synchronous calls to vehicle-service, spot-service,
    tariff-service and ticket-service.

    EVENT-DRIVEN ARCHITECTURE: not applicable in Phase 1. All inter-service
    communication is synchronous REST (output ports StayVehiclePort,
    StaySpotPort, StayTariffPort, StayTicketPort). Async messaging (RabbitMQ)
    and domain events are deferred to Phase 2.

    Stay STATE MACHINE (StayStatus):
      IN_PROGRESS -> FINISHED         (normal check-out, single call)
      IN_PROGRESS -> CANCELLED        (reverse-out at entry, CB06)
    FINISHED and CANCELLED are terminal and immutable (IN-19).
    NOTE: check-out is a single endpoint in Phase 1. Payment confirmation
    (RN-03) and the 10-min window (RN-04) are handled INTERNALLY in the
    CheckOutUseCase (mocked gateway), not as separate HTTP steps. The
    PAY_PENDING / PAID states remain in the domain enum for future use but are
    transient within a single check-out call.

    GAP FIXED: the CANCELLED state (CB06 - vehicle reverses out before
    crossing the entry security loop) was documented in StayStatus but had
    no reachable endpoint. POST /stays/{stayId}/cancel below covers the
    manual admin fallback described in CB06 for when the entry sensor fails
    to detect the reverse-out automatically. Confirm with the team whether
    an automatic sensor-triggered path (event/webhook) is also expected in
    Phase 1, or if manual-only is sufficient for now.
  version: 2.1.0
  contact:
    name: TARTIS Recon-AI - Team Parking

servers:
  - url: /v1

tags:
  - name: stays
    description: Stay lifecycle (entry, exit)

paths:
  /stays/check-in:
    post:
      tags: [stays]
      operationId: checkIn
      summary: Registers a vehicle entry (IN_PROGRESS)
      description: >
        Checks availability by vehicle type (RN-01), verifies the vehicle is not
        deactivated (RN-11) and has no other active stay (IN-02, CB05), assigns a
        spot, creates the stay and issues an entry ticket with barcode.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckInRequest'
      responses:
        '201':
          description: Stay created, spot assigned, entry ticket issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Vehicle not registered and insufficient data for auto-registration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: >
            No spots for the vehicle type (RN-01) or the vehicle already has an
            IN_PROGRESS stay (IN-02, CB05)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Vehicle deactivated (active=false), entry denied (RN-11)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  /stays/check-out:
    post:
      tags: [stays]
      operationId: checkOut
      summary: Full check-out in a single call (IN_PROGRESS -> FINISHED)
      description: >
        Triggered when the plate is read at the exit barrier. Locates the active
        stay by plate or entry ticket (HU-02 CA-01). In one atomic operation:
        calculates the amount via StayTariffPort applying minimum charge and
        courtesy (RN-09) and tiered pricing (RN-07, RN-08); verifies payment
        internally (RN-03, mocked in Phase 1); runs the mandatory sequence
        StayTicketPort (generates exit ticket) -> StaySpotPort (releases spot,
        RN-05); stamps checkOut and closes the stay. The entry ticket is
        invalidated (IN-21, not reusable).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckOutRequest'
      responses:
        '200':
          description: Check-out completed, exit ticket generated, spot released
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckOutResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: No active stay for the given plate/ticket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: >
            Stay is not IN_PROGRESS, or entry ticket already used (IN-21)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  /stays/{stayId}/cancel:
    post:
      tags: [stays]
      operationId: cancelStay
      summary: >
        Cancels a stay that never actually entered the parking (CB06:
        vehicle reversed out before crossing the entry security loop).
        Manual admin operation (RN-02) — fallback for when the entry
        sensor fails to detect the reverse-out automatically.
      description: >
        Releases the assigned spot (StaySpotPort) and invalidates the entry
        ticket, then transitions the stay to CANCELLED (IN-19, terminal and
        immutable).
      parameters:
        - $ref: '#/components/parameters/StayIdPath'
      responses:
        '200':
          description: Stay cancelled, spot released
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StayResponse'
        '404':
          description: Stay not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Stay is not IN_PROGRESS (already FINISHED or CANCELLED, IN-19)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  /stays/{stayId}:
    get:
      tags: [stays]
      operationId: getStay
      summary: Retrieves a stay detail (HU-08)
      parameters:
        - $ref: '#/components/parameters/StayIdPath'
      responses:
        '200':
          description: Stay detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StayResponse'
        '404':
          description: Stay not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  /stays:
    get:
      tags: [stays]
      operationId: listStays
      summary: Lists stays with filters (HU-08, admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/StayStatus'
        - name: plate
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Page of stays
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StayPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  schemas:

    VehicleType:
      type: string
      description: Canonical enum from the domain model
      enum: [CAR, CAR_PMR, MOTORBIKE]

    StayStatus:
      type: string
      description: >
        Stay states. PAID replaces the PIAD of the domain model (typo).
        CANCELLED covers IN-19 and CB06 (reverse-out at entry), not present in
        the domain model enum. PAY_PENDING / PAID are transient in Phase 1
        (single check-out endpoint) and kept for Phase 2.
      enum: [IN_PROGRESS, PAY_PENDING, PAID, FINISHED, CANCELLED]

    CheckInRequest:
      type: object
      required: [plate]
      properties:
        plate:
          type: string
          example: "1234ABC"
        vehicleType:
          allOf:
            - $ref: '#/components/schemas/VehicleType'
          description: >
            Required only if the vehicle is not registered (auto-registration).
            If it already exists, it is resolved via StayVehiclePort.

    EntryTicket:
      type: object
      description: Entry ticket (Entry_Ticket entity), issued at check-in
      properties:
        ticketId:
          type: string
          format: uuid
        barCode:
          type: string
          description: Barcode for exit-by-ticket (plate fallback)
        issuedAt:
          type: string
          format: date-time

    CheckInResponse:
      type: object
      properties:
        stayId:
          type: string
          format: uuid
        plate:
          type: string
        spotId:
          type: string
          format: uuid
        checkIn:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/StayStatus'
        entryTicket:
          $ref: '#/components/schemas/EntryTicket'

    CheckOutRequest:
      type: object
      description: >
        Either plate OR entryTicketId must be provided (at least one).
        entryTicketId is the fallback when the plate cannot be read
        (HU-02 CA-01, CB10).
      properties:
        plate:
          type: string
          example: "1234ABC"
        entryTicketId:
          type: string
          format: uuid

    CheckOutResponse:
      type: object
      properties:
        stayId:
          type: string
          format: uuid
        plate:
          type: string
        checkIn:
          type: string
          format: date-time
        checkOut:
          type: string
          format: date-time
        totalMinutes:
          type: integer
          description: Minutes rounded up in the system's favor (RN-06)
        amount:
          type: number
          format: double
        ticketId:
          type: string
          format: uuid
          description: Exit ticket (Ticket entity / payment receipt)
        status:
          $ref: '#/components/schemas/StayStatus'

    StayResponse:
      type: object
      properties:
        stayId:
          type: string
          format: uuid
        plate:
          type: string
        vehicleId:
          type: string
          format: uuid
        spotId:
          type: string
          format: uuid
        tariffId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/StayStatus'
        checkIn:
          type: string
          format: date-time
        checkOut:
          type: string
          format: date-time
          nullable: true
          description: null if and only if status = IN_PROGRESS (IN-14)
        totalAmount:
          type: number
          format: double
          nullable: true

    StayPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/StayResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 409
        error:
          type: string
          example: "CONFLICT"
        message:
          type: string
          example: "No available spots for vehicle type CAR"
        path:
          type: string
          example: "/v1/stays/check-in"

  parameters:
    StayIdPath:
      name: stayId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    BadRequest:
      description: Invalid request (field validation)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT